📝 docs(wezterm): add commented alternative wezterm config method
For if you'd prefer not to use submodules, these commented blocks could be used to have chezmoi handle the git ops.

This is just for reference. It has no impact on the normal dotfiles configuration, and in fact would likely require some manual editing to use instead. But it's here if you need it.